added geometry loader to core group for level 2
